Update demo application
